requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
tqdm
matplotlib
seaborn
PyYAML
transformers
huggingface_hub
